fix: cast timestep to float in PredictionHeadWrapper for ONNX export
The Gemm node in prediction_head required float inputs but timestep
was int64, causing OnnxRuntimeException at model load time.
Added .float() cast in the wrapper's forward method and re-exported.

+++ src/scenario-08-onnx-native/export/export_model.py
@@ -75,7 +75,8 @@ class PredictionHeadWrapper(nn.Module):
 
     def forward(self, noisy_latent: torch.Tensor, timestep: torch.Tensor,
                 conditioning: torch.Tensor) -> torch.Tensor:
-        return self.head(noisy_latent, timestep, conditioning)
+        # Cast timestep to float — internal linear layers expect float, not int64
+        return self.head(noisy_latent, timestep.float(), conditioning)
 
 
 class AcousticDecoderWrapper(nn.Module):
